video_processor/requirements.txt
boto3
google-generativeai
opencv-python-headless
requests
